requirements.txt
Django>=2.0.2
django-hijack
django-hijack-admin
django-extensions
jupyter
django-debug-toolbar
bootstrap4